smoother transition added
